requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
altair==4.2.2
    # via streamlit
async-timeout==4.0.2
    # via redis
attrs==22.2.0
    # via jsonschema
blinker==1.5
    # via streamlit
cachetools==5.3.0
    # via streamlit
certifi==2022.12.7
    # via requests
charset-normalizer==3.0.1
    # via requests
click==8.1.3
    # via streamlit
decorator==5.1.1
    # via validators
entrypoints==0.4
    # via altair
gitdb==4.0.10
    # via gitpython
gitpython==3.1.30
    # via streamlit
idna==3.4
    # via requests
importlib-metadata==6.0.0
    # via streamlit
jinja2==3.1.2
    # via
    #   altair
    #   pydeck
jsonschema==4.17.3
    # via altair
markdown-it-py==2.1.0
    # via rich
markupsafe==2.1.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
numpy==1.24.1
    # via
    #   altair
    #   pandas
    #   pyarrow
    #   pydeck
    #   scipy
    #   streamlit
packaging==23.0
    # via streamlit
pandas==1.5.3
    # via
    #   altair
    #   streamlit
    #   streamlit-aggrid
pillow==9.4.0
    # via streamlit
plotly==5.13.0
    # via -r requirements.in
protobuf==3.20.3
    # via streamlit
pyarrow==11.0.0
    # via streamlit
pydeck==0.8.0
    # via streamlit
pygments==2.14.0
    # via rich
pympler==1.0.1
    # via streamlit
pyrsistent==0.19.3
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   pandas
    #   streamlit
python-decouple==3.7
    # via streamlit-aggrid
pytz==2022.7.1
    # via pandas
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
redis==4.4.2
    # via spotipy
requests==2.28.2
    # via
    #   spotipy
    #   streamlit
rich==13.3.0
    # via streamlit
scipy==1.10.0
    # via -r requirements.in
semver==2.13.0
    # via streamlit
six==1.16.0
    # via
    #   python-dateutil
    #   spotipy
smmap==5.0.0
    # via gitdb
spotipy==2.22.1
    # via -r requirements.in
streamlit==1.17.0
    # via
    #   -r requirements.in
    #   streamlit-aggrid
streamlit-aggrid==0.3.3
    # via -r requirements.in
tenacity==8.1.0
    # via plotly
toml==0.10.2
    # via streamlit
toolz==0.12.0
    # via altair
tornado==6.2
    # via streamlit
typing-extensions==4.4.0
    # via streamlit
tzdata==2022.7
    # via pytz-deprecation-shim
tzlocal==4.2
    # via streamlit
urllib3==1.26.14
    # via
    #   requests
    #   spotipy
validators==0.20.0
    # via streamlit
zipp==3.12.0
    # via importlib-metadata
